+++ app/src/main/java/com/sanha/moneytalk/core/firebase/PremiumManager.kt
@@ -299,7 +299,7 @@ class PremiumManager @Inject constructor(
             maintenanceMessage = snapshot.child("maintenance_message").getValue(String::class.java) ?: "",
             rewardAdEnabled = snapshot.child("reward_ad_enabled").getValue(Boolean::class.java) ?: false,
             creditAdEnabled = snapshot.child("credit_ad_enable").getValue(Boolean::class.java) ?: false,
-            rewardAdChatCount = snapshot.child("reward_ad_chat_count").getValue(Int::class.java) ?: 5,
+            rewardAdChatCount = snapshot.child("reward_ad_chat_count").getValue(Int::class.java) ?: 2,
             sendOriginMessage = snapshot.child("send_origin_message").getValue(Boolean::class.java) ?: false,
             freeSyncCount = snapshot.child("free_sync_count").getValue(Int::class.java) ?: 3,
             minVersionCode = snapshot.child("min_version_code").getValue(Int::class.java) ?: 1,